requirements.txt
absl-py==1.0.0
attrs==21.4.0
certifi==2022.5.18.1
cycler==0.11.0
dataclasses==0.6
fonttools==4.33.3
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
mediapipe==0.8.3.1
numpy==1.21.6
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.3.5
Pillow==9.1.1
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
torch==1.11.0
typing_extensions==4.2.0
